Update App so AppRoutes enforces all routing, including authentication
